Cleanup EventQuery
